Refactor hexagonal  | moved bots
