+++ src/bot/domain/dto.py
@@ -1,0 +1,6 @@
+import typing as t
+
+
+class ChatBotReply(t.TypedDict):
+    reply: str
+    context: list[dict]
